backend/requirements.txt
fastapi
uvicorn
spotipy
langchain
langgraph
langchain-google-genai
pedalboard
python-dotenv
requests
youtube-search-python
